+++ tests/test_column_aliases.py
@@ -1,4 +1,5 @@
 import sqlalchemy as sa
+from sqlalchemy_continuum import version_class
 
 from tests import TestCase, create_test_cases
 
@@ -9,7 +10,9 @@ class ColumnAliasesTestCase(TestCase):
             __tablename__ = 'text_item'
             __versioned__ = {}
 
-            id = sa.Column(sa.Integer, autoincrement=True, primary_key=True)
+            id = sa.Column(
+                '_id', sa.Integer, autoincrement=True, primary_key=True
+            )
 
             name = sa.Column('_name', sa.Unicode(255))
 
@@ -29,6 +32,27 @@ class ColumnAliasesTestCase(TestCase):
         self.session.commit()
         item.versions[0].revert()
         self.session.commit()
+
+    def test_previous_for_deleted_parent(self):
+        item = self.TextItem()
+        item.name = u'Some item'
+        item.content = u'Some content'
+        self.session.add(item)
+        self.session.commit()
+        self.session.delete(item)
+        self.session.commit()
+        TextItemVersion = version_class(self.TextItem)
+
+        versions = (
+            self.session.query(TextItemVersion)
+            .order_by(
+                getattr(
+                    TextItemVersion,
+                    self.options['transaction_column_name']
+                )
+            )
+        ).all()
+        assert versions[1].previous.name == u'Some item'
 
 
 create_test_cases(ColumnAliasesTestCase)